+++ package.json
@@ -51,22 +51,25 @@
         "@babel/runtime": "^7.12.5",
         "@energyweb/iam-contracts": "1.17.2",
         "@ensdomains/ens": "^0.4.5",
-        "@ew-did-registry/claims": "0.5.2-alpha.1076.0",
-        "@ew-did-registry/did": "0.5.2-alpha.1076.0",
-        "@ew-did-registry/did-document": "0.5.2-alpha.1076.0",
-        "@ew-did-registry/did-ethr-resolver": "0.5.2-alpha.1076.0",
-        "@ew-did-registry/did-resolver-interface": "0.5.2-alpha.1076.0",
-        "@ew-did-registry/jwt": "0.5.2-alpha.1076.0",
-        "@ew-did-registry/keys": "0.5.2-alpha.1076.0",
-        "@ew-did-registry/proxyidentity": "0.5.2-alpha.1076.0",
+        "@ew-did-registry/claims": "^0.5.2-alpha.65.0",
+        "@ew-did-registry/did": "^0.5.2-alpha.65.0",
+        "@ew-did-registry/did-document": "^0.5.2-alpha.65.0",
+        "@ew-did-registry/did-ethr-resolver": "^0.5.2-alpha.65.0",
+        "@ew-did-registry/did-resolver-interface": "^0.5.2-alpha.65.0",
+        "@ew-did-registry/jwt": "^0.5.2-alpha.65.0",
+        "@ew-did-registry/keys": "^0.5.2-alpha.65.0",
+        "@ew-did-registry/proxyidentity": "^0.5.2-alpha.65.0",
         "@gnosis.pm/safe-apps-sdk": "1.0.3",
         "@metamask/detect-provider": "^1.2.0",
+        "@nestjs/common": "^8.0.6",
+        "@nestjs/config": "^1.0.1",
+        "@nestjs/core": "^8.0.6",
         "@walletconnect/web3-provider": "1.0.0-rc.4",
         "axios": "^0.21.1",
         "base64url": "^3.0.1",
         "concurrently": "^6.1.0",
         "eth-ens-namehash": "^2.0.8",
-        "ethers": "4.0.45",
+        "ethers": "^5.4.6",
         "js-sha3": "^0.8.0",
         "lodash.difference": "^4.5.0",
         "nats.ws": "1.0.0-111",
