+++ src/config/chain.config.ts
@@ -8,10 +8,8 @@ import {
     VOLTA_CLAIM_MANAGER_ADDRESS,
     VOLTA_STAKING_POOL_FACTORY_ADDRESS,
 } from "@energyweb/iam-contracts";
-import { CacheServerClientOptions } from "../cacheServerClient/cacheServerClient";
-import { MessagingMethod } from "../utils/constants";
+import { VOLTA_CHAIN_ID } from "../constants";
 
-const VOLTA_CHAIN_ID = 73799;
 export interface ChainConfig {
     rpcUrl: string;
     ensRegistryAddress: string;
@@ -24,16 +22,11 @@ export interface ChainConfig {
     stakingPoolFactoryAddress: string;
 }
 
-export interface MessagingOptions {
-    messagingMethod: MessagingMethod;
-    natsServerUrl: string;
-}
-
 /**
  * Set of parameters to configure connection to chain with id received from wallet.
  * If configuration for some chain is missing or should be reconfigured use `setChainConfig` before class instantiation
  */
-export const chainConfigs: Record<number, ChainConfig> = {
+const chainConfig: Record<number, ChainConfig> = {
     [VOLTA_CHAIN_ID]: {
         rpcUrl: "https://volta-rpc.energyweb.org",
         ensRegistryAddress: VOLTA_ENS_REGISTRY_ADDRESS,
@@ -47,41 +40,12 @@ export const chainConfigs: Record<number, ChainConfig> = {
     },
 };
 
-export const cacheServerClientOptions: Record<number, CacheServerClientOptions> = {
-    [VOLTA_CHAIN_ID]: {
-        url: "https://volta-identitycache.energyweb.org/",
-        cacheServerSupportsAuth: true,
-    },
-};
-
-export const messagingOptions: Record<number, MessagingOptions> = {
-    [VOLTA_CHAIN_ID]: {
-        messagingMethod: MessagingMethod.Nats,
-        natsServerUrl: "https://volta-identityevents.energyweb.org/",
-    },
-};
+export default () => ({ chainConfig });
 
 /**
  * Used to override existing chain configuration or add a missing one
  * Configuration must be set before constructing `IAM`
  */
 export const setChainConfig = (chainId: number, config: Partial<ChainConfig>) => {
-    chainConfigs[chainId] = { ...chainConfigs[chainId], ...config };
-};
-
-/**
- * Used to override existing cache server configuration or add a missing one
- * Configuration must be set before constructing `IAM`
- */
-export const setCacheClientOptions = (chainId: number, options: Partial<CacheServerClientOptions>) => {
-    cacheServerClientOptions[chainId] = { ...cacheServerClientOptions[chainId], ...options };
-};
-
-/**
- * Used to override existing messaging configuration or add a missing one
- * Configuration must be set before constructing `IAM`
- */
-
-export const setMessagingOptions = (chainId: number, options: Partial<MessagingOptions>) => {
-    messagingOptions[chainId] = { ...messagingOptions[chainId], ...options };
+    chainConfig[chainId] = { ...chainConfig[chainId], ...config };
 };
