+++ src/constants.ts
@@ -1,0 +1,1 @@
+export const VOLTA_CHAIN_ID = 73799;
