+++ src/modules/assets/assets.module.ts
@@ -1,0 +1,7 @@
+import { Module } from "@nestjs/common";
+
+@Module({
+    providers: [AssetsService],
+    exports: [AssetsService],
+})
+export class AssetsModule {}
