+++ src/modules/domains/domains.types.ts
@@ -1,0 +1,5 @@
+export enum ENSNamespaceTypes {
+    Roles = "roles",
+    Application = "apps",
+    Organization = "org",
+}
